requirements.txt
Django==5.2.6
djangorestframework==3.16.1
Pillow==11.3.0
django-extensions==4.1